[WA-4653] change program and org ID type
* we changed the program and organization ID type to bigint to
accomodate a 64 bit integer coming from CAS, that will be used as a
unique identifier for those models across applicantions and
environments. the identifier will remain unchanged in CAS 3 prod as
well as UAT envs, and in WA envs.
* cas 2 will continue to use the table PK as the id; the PK will be
copied to the new field to keep the API standard for all.