[NF-39394] Add the documentation for batch designations
We're adding the new documentation for the batch designations endpoint.
We want that doc to be available to make it easier to review the
documentation internally. But we don't want to make it visible for the
public as it's not done yet. That's why we're omitting the link in the
menu for now. The page will be available with the direct URL but not
trough the menu.